requirements.txt
phidata
python-dotenv
chainlit
PyMySQL
cryptography
sqlalchemy
ollama
